requirements.txt
python-docx
dictdiffer